abstract enemy classes
created a base class for enemies to inherit from and commented code
