+++ BingerNinja/Assets/Scripts/AI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5752c3b1ae00af24abb42774b70c4467
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
