+++ BingerNinja/Assets/Scripts/AI/MeleeEnemy_SebastianMol.cs
@@ -1,0 +1,22 @@
+﻿//sebastian mol
+//melle enemy class can aslo be used for brute enemy
+class MeleeEnemy_SebastianMol : BaseEnemy_SebastianMol
+{
+    public override void MovementBehaviour()
+    {
+        throw new System.NotImplementedException();
+    }
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+
+}
