+++ BingerNinja/Assets/Scripts/AI/MeleeEnemy_SebastianMol.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 131872bffc30dbd4bb3013b65a7f2c34
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
