+++ BingerNinja/Assets/Scripts/AI/RangedEnemy_SebastianMol.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8e6d75813f5b8f543aa3dbdb51792783
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
