+++ BingerNinja/Assets/Scripts/AI/WhimpEnemy_SebastianMol.cs
@@ -1,0 +1,23 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+class WhimpEnemy_SebastianMol : BaseEnemy_SebastianMol
+{
+    public override void MovementBehaviour()
+    {
+        throw new System.NotImplementedException();
+    }
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+}
