+++ BingerNinja/Assets/Scripts/AI/WhimpEnemy_SebastianMol.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f70c1f3b20b41b149aef798f8ebb16bf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
